% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{c19}
\alias{c19}
\alias{c19l}
\title{Covid-19 Datasets}
\format{
An object of class \code{data.frame} with 14985 rows and 6 columns.

An object of class \code{data.frame} with 49356 rows and 4 columns.
}
\source{
\url{https://raw.githubusercontent.com/CSSEGISandData/COVID-19/}
}
\usage{
c19 # wide format

c19l # long format
}
\description{
Covid-19 datasets various resources.

\code{c19}: John-Hopkins University
}
\details{
Some countries which has provinces were seperated from the original data source.
These countries are Autralia, Canada, China, Denmark, France, United Kingdom
and Netherlands. You can see long format of these data sources by adding
3-letters iso code at the end of 'c19.' string i.e., 'c19.gbr'.

\code{[dataset_name]l}: long format
}
\keyword{datasets}
