% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{c19jhw}
\alias{c19jhw}
\alias{c19jhl}
\title{Covid-19 Datasets}
\format{
An object of class \code{data.frame} with 19462 rows and 8 columns.

An object of class \code{data.frame} with 56462 rows and 7 columns.
}
\source{
\url{https://raw.githubusercontent.com/CSSEGISandData/COVID-19/}
}
\usage{
c19jhw # wide format

c19jhl # long format
}
\description{
Covid-19 datasets various resources.

\code{c19jh}: John-Hopkins University
}
\details{
\code{[dataset_name]w}: wide format

\code{[dataset_name]l}: long format
}
\keyword{datasets}
