% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{lookup}
\alias{lookup}
\title{Lookup Table}
\format{
An object of class \code{data.frame} with 3590 rows and 13 columns.
}
\source{
\url{https://raw.githubusercontent.com/CSSEGISandData/COVID-19/}
}
\usage{
lookup
}
\description{
Lookup Table
}
\keyword{datasets}
