% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{Yearlycumulative_vaccine}
\alias{Yearlycumulative_vaccine}
\title{Function for returning data frame for the Covid - 19 vaccination which is cumulative in different provinces in Canada.}
\usage{
Yearlycumulative_vaccine(provinceName = "Canada", cumulativeyear = "2020")
}
\arguments{
\item{provinceName}{a character/ string depicting the name of the province}

\item{cumulativeyear}{a character/ string for the year}
}
\value{
Data frame for the Covid - 19 cummulative vaccines corresponding to a particular province and year
}
\description{
Performed data wrangling and cleaning using the API for the Covid - 19 cases in Canada.
It processes the API and returns the data corresponding to one province
which is passed on as the argument. If user passes empty argument, so by default Canada is used which returns the data of whole Canada as a whole.
The returned data is a data frame and contains the columns including the date of the vaccination, province name,cumulative vaccine and cvaccine.
}
\examples{
Yearlycumulative_vaccine('Alberta','2020')
}
