% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{c19dw}
\alias{c19dw}
\alias{c19ra}
\alias{c19jh}
\alias{c19trl}
\alias{c19trw}
\title{Covid-19 dataset}
\format{
An object of class \code{data.frame} with 54712 rows and 7 columns.

An object of class \code{data.frame} with 54351 rows and 7 columns.

An object of class \code{data.frame} with 15591 rows and 7 columns.

An object of class \code{data.frame} with 48 rows and 7 columns.

An object of class \code{data.frame} with 12 rows and 5 columns.
}
\source{
\code{c19dw}: \url{https://query.data.world}

\code{c19ra}: \url{https://raw.githubusercontent.com/RamiKrispin/coronavirus-csv/master/coronavirus_dataset.csv}

\code{c19jh}: \url{https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports/}

\code{c19tr}: \url{https://twitter.com/drfahrettinkoca}

\code{c19tr}: \url{https://twitter.com/drfahrettinkoca}
}
\description{
\code{c19dw}: Data from data.world

\code{c19ra}: Data from RamiKrispin GitHub repository

\code{c19jh}: Data from John Hopkins University

\code{c19trl}: Data for Turkey (Long format)

\code{c19trl}: Data for Turkey (Wide format)
}
\details{
Covid-19 datasets from various sources
}
\keyword{datasets}
