% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{c19jhw}
\alias{c19jhw}
\alias{c19jhl}
\title{Covid-19 dataset}
\format{
An object of class \code{data.frame} with 19126 rows and 8 columns.

An object of class \code{data.frame} with 55480 rows and 7 columns.
}
\source{
\code{c19jh}: \url{https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports/}
}
\description{
\code{c19jhw}: Data from John Hopkins University (Wide format)

\code{c19jhl}: Data from John Hopkins University (Long format)
}
\details{
Covid-19 datasets from various sources

[datasetname]w is wide format

[datasetname]l is long format
}
\keyword{datasets}
