% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{c19dw_w}
\alias{c19dw_w}
\alias{c19dw_l}
\alias{c19ra_w}
\alias{c19ra_l}
\alias{c19jh_w}
\alias{c19jh_l}
\alias{c19tr_w}
\alias{c19tr_l}
\title{Covid-19 dataset}
\format{
An object of class \code{data.frame} with 17690 rows and 7 columns.

An object of class \code{data.frame} with 35380 rows and 7 columns.

An object of class \code{data.frame} with 14756 rows and 7 columns.

An object of class \code{data.frame} with 29512 rows and 7 columns.

An object of class \code{data.frame} with 12021 rows and 11 columns.

An object of class \code{data.frame} with 48084 rows and 9 columns.

An object of class \code{data.frame} with 14 rows and 5 columns.

An object of class \code{data.frame} with 56 rows and 7 columns.
}
\source{
\code{c19dw}: \url{https://query.data.world}

\code{c19ra}: \url{https://raw.githubusercontent.com/RamiKrispin/coronavirus-csv/master/coronavirus_dataset.csv}

\code{c19ra}: \url{https://raw.githubusercontent.com/RamiKrispin/coronavirus-csv/master/coronavirus_dataset.csv}

\code{c19jh}: \url{https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports/}

\code{c19tr}: \url{https://twitter.com/drfahrettinkoca}
}
\description{
\code{c19dw_w}: Data from data.world (Wide format)

\code{c19dw_l}: Data from data.world (Long format)

\code{c19ra_w}: Data from RamiKrispin GitHub repository (Wide format)

\code{c19ra_l}: Data from RamiKrispin GitHub repository (Long format)

\code{c19jh_w}: Data from John Hopkins University (Wide format)

\code{c19jh_l}: Data from John Hopkins University (Long format)

\code{c19tr_w}: Data for Turkey (Wide format)

\code{c19tr_l}: Data for Turkey (Long format)
}
\details{
Covid-19 datasets from various sources

[dataset]_w is wide format

[dataset]_l is long format
}
\keyword{datasets}
