% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{c19dww}
\alias{c19dww}
\alias{c19dwl}
\alias{c19raw}
\alias{c19ral}
\alias{c19jhw}
\alias{c19jhl}
\title{Covid-19 dataset}
\format{
An object of class \code{data.frame} with 20553 rows and 7 columns.

An object of class \code{data.frame} with 41100 rows and 7 columns.

An object of class \code{data.frame} with 16951 rows and 7 columns.

An object of class \code{data.frame} with 33902 rows and 7 columns.

An object of class \code{data.frame} with 17219 rows and 8 columns.

An object of class \code{data.frame} with 50660 rows and 7 columns.
}
\source{
\code{c19dw}: \url{https://query.data.world}

\code{c19ra}: \url{https://raw.githubusercontent.com/RamiKrispin/coronavirus-csv/master/coronavirus_dataset.csv}

\code{c19ra}: \url{https://raw.githubusercontent.com/RamiKrispin/coronavirus-csv/master/coronavirus_dataset.csv}

\code{c19jh}: \url{https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports/}
}
\description{
\code{c19dww}: Data from data.world (Wide format)

\code{c19dwl}: Data from data.world (Long format)

\code{c19raw}: Data from RamiKrispin GitHub repository (Wide format)

\code{c19ral}: Data from RamiKrispin GitHub repository (Long format)

\code{c19jhw}: Data from John Hopkins University (Wide format)

\code{c19jhl}: Data from John Hopkins University (Long format)
}
\details{
Covid-19 datasets from various sources

[dataset]w is wide format

[dataset]l is long format
}
\keyword{datasets}
